Revert fix of string-escapes
The problem is, that baselib in ckl-js is provided as a
string in javascript. This makes string escapes complicated.
It would be better to read a text file, as it is done in
the java and the dotnet versions. But this must work both
in the browser as well as with node.js.