Fix bugs and make functionality uniform on all platforms.
Basically, this was done by adding the benchmark originally
written for the dotnet platform and running it on the other
platforms. This resulted in a variety of errors and by fixing
them, we now have some bugs less in the js and java platforms.

Also interesting are the performance differences, and a bit
surprising is that the dotnet platform has worst performance!
To be fair, it is still run in debug mode, maybe this does
make a bit of a difference.