Tweak handling of value comparison for string, int and decimal
This issue is surprisingly difficult. Now, the approach is to
check whether both values are numerical. If this is the case,
then we compare the values (after potentially converting an
int to a decimal, if one of the values is a decimal). Otherwise
both values are converted to their string representation and
this representation is compared.

I think this should now work correctly. The other values already
converted to string representation and compared them.