Implement basic module system in ckl-js
This adds a very simple module system to
Checkerlang. It is usable in the node-js
version of ckl-js.

Basically, one can define modules and add
them to a program using a new "require"
statement. This loads the module and provides
it or its symbols in the current environment.

This still has many limitations and is not
fully implemented. For example, the location
of system provided modules has to be handled
correctly. Also, in the browser-based version,
the system modules have to be implemented
within the baselib module, probably.

Also, the porting of this system to ckl-java
and ckl-dotnet is not done yet.

Finally, the baselib has to be split into
logical modules. Also, a backwards compatible
mode has to be available so that older programs
work unchanged.