Better handle module dependencies
Currently, modules are expected to form a tree
and not contain circles.

The problem is, that at the location of a require
the module is loaded and all its symbols are
added. If this module itself contains a require
for the current module, then the first is already
(partially) loaded and does probably not contain
all required definitions (e.g. everything after
the original require statement is missing). Thus
this does not really work correctly.

Better would be the following approach when
requiring a module: first parse and evaluate all
non-require statements and expressions. Then, in
a second pass, process all require statements.

This works fine for function definitions, but not
for circular variable definitions (i.e. if a def
depends on a variable in a separate module, then
this variable is just not yet there). But maybe
this is a price we have to pay. The modules are
mainly intended for functions and less for constants.