Further tweak module mechanism by detecting circular dependencies
I decided to disallow circular dependencies. Because of this, I
could revert the two last changes, namely the require statement
is now not required to be top-level anymore and no reordering
of the require statements to the end of the top-level block occurs
anymore.

Instead, we track the current require-stack and detect circles.
Thus, we can abort and provide quite good error messages.

This simplifies the system and makes it more intuitive and more
usable, I think.