Implement js module loader and split off math functions
This is kind of the proof of concept for the modules.
Now, there is a math module which contains many of the
math functions. For backwards compatibility, this module
is required unqualified in the base module (formerly
baselib), thus all functions are available in the global
environment as before.

Later on, more modules are added and not all modules are
imported into the global environment. Thus, to use e.g.
math functions like sin, it will be required to import
the math module either unqalified or qualified.