Do not open a new scope in for construct
The only variable scopes in Checkerlang are opened
by functions (and in a limited way by the list
comprehension construct). Blocks generally do not
create a new environment.

In the for statements, an earlier change introduced
a new scope for the iteration variable and consequently
for all code within the for loop body.

This change reverts that.

A remaining problem is what to do if the for loop
iteration variable already exists. The current
implementation overwrites the variable, and removes
it at the end of the for loop. Thus, the old variable
is then not visible anymore.

There are two solutions for this: either throw an error
if the variable already exists, or save the old values
and restore them at the end of the for loop. The latter
would be more smooth, the former more principled. Currently
neither is implemented. This needs to be done in an later
change.