Add legacy and modern mode to notebook, provide more scripts
Now it is possible to switch between modern and legacy modes
of the interpreter. Also, instead of two scripts, five are
available, albeit with fixed names A, B, C, D and E. Reloading
a script clear the output to make it clear that the interpreter
is in a newly initialized state.