Rework secure mode by handling it in bind_native.
This makes a lot of stuff easier and removes the need
to remove the bind_native function from the environment
in case of secure mode. We need the bind_native function
when loading modules that use it.

Now, the bind_native checks the secure flag and each
native function has a flag whether it is secure or
not. In secure mode, only secure function may be bound
using bind_native.