Add parallel and product list comprehensions
These are variants of list comprehensions which
handle two lists and process them either in
parallel or as a cartesian product. More than
two lists are not allowed.

This also cleans up the code a little bit by
extracting collection accessing code into a
function.

Please be aware that this is only implements in
ckl-js as of yet. The other platforms will follow.

Also: these variants only exist in list comprehensions
neither in set nor map comprehensions. They will
be added later on.