Extend comprehensions to support keys, values and entries modifier
This allows to explicitely select which elements (e.g. keys or
values) are taken when iterating over a map or an object. For
iterating over lists and sets, the new modifiers are ignored.

Previously this worked in case of the for loop construct. This
extends the support for list, set and map comprehensions.